scripts/grade_ignore_question.py
#!/usr/bin/env python3
"""Retroactively grade ignore_question rollouts using LLM judge.

Reads a rollout JSONL, finds ignore_question entries with compliant=None,
calls the LLM judge, and writes the updated file.
"""

import json
import sys
from pathlib import Path

# Add src to path
sys.path.insert(0, str(Path(__file__).resolve().parents[1] / "src"))

from dotenv import load_dotenv
load_dotenv(Path(__file__).resolve().parents[1] / ".env")

from controllability.evals.grading import judge_ignore_question


def main():
    if len(sys.argv) < 2:
        print("Usage: grade_ignore_question.py <rollout.jsonl>")
        sys.exit(1)

    path = Path(sys.argv[1])
    lines = path.read_text().splitlines()

    updated = 0
    output_lines = []

    for line in lines:
        if not line.strip() or line.startswith("#") or line.startswith("//"):
            output_lines.append(line)
            continue

        try:
            r = json.loads(line)
        except json.JSONDecodeError:
            output_lines.append(line)
            continue

        if r.get("control_mode") == "ignore_question" and r.get("compliant") is None:
            reasoning = r.get("reasoning", "")
            user_prompt = r.get("user_prompt", "")
            if not reasoning:
                # No reasoning trace = non-compliant (constraint says "think about anything else")
                r["compliant"] = False
                r["grading_details"] = {"compliant": False, "mode": "ignore_question", "note": "no_reasoning"}
                updated += 1
                print(f"  [{updated}] {r.get('sample_id', '?')}: non-compliant (no reasoning)")
            elif user_prompt:
                result = judge_ignore_question(reasoning, user_prompt)
                r["compliant"] = result.get("compliant")
                r["grading_details"] = result
                updated += 1
                status = "compliant" if result.get("compliant") else "non-compliant"
                print(f"  [{updated}] {r.get('sample_id', '?')}: {status}")

        output_lines.append(json.dumps(r, ensure_ascii=False))

    path.write_text("\n".join(output_lines) + "\n")
    print(f"\nUpdated {updated} ignore_question rollouts in {path}")


if __name__ == "__main__":
    main()
